fix: restructured folder layout and minor changes
