+++ src/control/ControlManager.ts
@@ -1,0 +1,1 @@
+export abstract class ControlManager {}
