--- src/plot/BarPlot.ts
@@ -1,11 +1,0 @@
-import { Position } from "../helpers/Position";
-import { Plot } from "./Plot";
-
-export class BarPlot extends Plot {
-  maxY: number = 0;
-  maxX: number = 0;
-  minY: number = 0;
-  minX: number = 0;
-  plotData: any;
-  position: Position | null = null;
-}
