+++ src/plots/bar/BarControl.ts
@@ -1,0 +1,3 @@
+import { ControlManager } from "../../control/ControlManager";
+
+class BarControl extends ControlManager {}
